+++ src/js/tracing/integrations/appStart.ts
@@ -22,6 +22,7 @@ import {
   APP_START_WARM as APP_START_WARM_OP,
   UI_LOAD as UI_LOAD_OP,
 } from '../ops';
+import { getReactNativeTracingIntegration } from '../reactnativetracing';
 // import { getReactNativeTracingIntegration } from '../reactnativetracing';
 import { SEMANTIC_ATTRIBUTE_SENTRY_OP } from '../semanticAttributes';
 import { createChildSpanJSON, createSpanJSON, getBundleStartTimestampMs } from '../utils';
@@ -108,7 +109,7 @@ export const appStartIntegration = ({
   standalone?: boolean;
 } = {}): AppStartIntegration => {
   let _client: Client | undefined = undefined;
-  const standalone = standaloneUserOption;
+  let standalone = standaloneUserOption;
   let isEnabled = true;
   let appStartDataFlushed = false;
 
@@ -123,10 +124,10 @@ export const appStartIntegration = ({
     }
   };
 
-  const afterAllSetup = (_client: Client): void => {
+  const afterAllSetup = (client: Client): void => {
     if (standaloneUserOption === undefined) {
       // If not user defined, set based on the routing instrumentation presence
-      // FIXME: standalone = getReactNativeTracingIntegration(client)?.options.routingInstrumentation;
+      standalone = !getReactNativeTracingIntegration(client)?.options.routingInstrumentation;
     }
   };